requirements.txt
python-chess
json
multiprocessing
requests